% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fn_v_cpp}
\alias{fn_v_cpp}
\title{Synth Optimization func for predictor variable weights}
\usage{
fn_v_cpp(variables_v, X0_scaled, X1_scaled, Z0, Z1)
}
\arguments{
\item{variables_v}{The weights on the Synth predictor variables}

\item{X0_scaled}{scaled matrix of control units}

\item{X1_scaled}{scaled matrix for the treated unit}

\item{Z0}{matrix of outcome observations for the control units}

\item{Z1}{matrix of outcome observations for the treated unit}
}
\value{
the loss from the parameter optimizaiton
}
\description{
The Optimization function for the synthetic control implemented
using Rcpp
}
\author{
Chandler Lutz
}

