% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solution.v}
\alias{solution.v}
\title{An example solution.v (weights on the variables) in the
for the synth class}
\format{vector}
\usage{
solution.v
}
\description{
An example solution.v (weights on the variables) in the
for the synth class
}
\keyword{datasets}

