% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solution_w_cpp}
\alias{solution_w_cpp}
\title{solution for region weights}
\usage{
solution_w_cpp(solution_v, X0_scaled, X1_scaled)
}
\arguments{
\item{solution_v}{the weighte on each predictor}

\item{X0_scaled}{scaled matrix of control units}

\item{X1_scaled}{scaled matrix for the treated unit}
}
\value{
A vector with the solution with the regional weights
}
\description{
To get the solution for the weights from the weights on each
predictor variable and the X scaled matrices for the treated unit
and control units
}
\author{
Chandler Lutz
}

