% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynthHelpers_main.R
\docType{package}
\name{SynthHelpers-package}
\alias{SynthHelpers}
\alias{SynthHelpers-package}
\title{Helpers for the R6 Synthetic Control class}
\description{
Use functions from the Synthetic Control class
}

