% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X.scaled}
\alias{X.scaled}
\title{An example of an X.scaled matrix of for each region
of input data}
\format{matrix}
\usage{
X.scaled
}
\description{
An example of an X.scaled matrix of for each region
of input data
}
\keyword{datasets}

